BipartiteIndexedGraph

A graph is bipartite if the set of vertices can be partitioned into two blocks such that there is no edge between vertices of different blocks. Here we adopt the notation of referring to these as the "left" and "right" block.

A BipartiteIndexedGraph behaves just like a bipartite, undirected IndexedGraph, with two differences:

Vertices can be indexed as usual via their integer index (which is called here a linearindex), or via a BipartiteGraphVertex, i.e. by specifying Left or Right and the integer index of that vertex within its block. The typical use case is that where one has two vectors storing vertex properties of the two blocks, possibly with different eltypes, and each with indices starting at one.

The adjacency matrix of a bipartite graph (possibly after permutation of the vertex indices) is made of two specular rectangular submatrices A. Only one of these is stored, leading to a slight improvement in efficiency.

BipartiteIndexedGraphs use the same edge type IndexedEdge as the other AbstractIndexedGraphs, which stores source and destination as linear indices. To retrieve the in-block indices of the two vertices incident on an edge, use vertex_left, vertex_right.

IndexedGraphs.BipartiteIndexedGraph — Type

BipartiteIndexedGraph{T<:Integer} <: AbstractIndexedGraph{T}

A type representing a sparse, undirected bipartite graph.

FIELDS

A – adjacency matrix filled with NullNumbers. Rows are vertices belonging to the left block, columns to the right block

X – square matrix for efficient access by row. X[j,i] points to the index of element A[i,j] in A.nzval.

IndexedGraphs.BipartiteIndexedGraph — Method

BipartiteIndexedGraph(A::AbstractMatrix)

Construct a BipartiteIndexedGraph from adjacency matrix A with the convention that rows are vertices belonging to the left block, columns to the right block

IndexedGraphs.BipartiteIndexedGraph — Method

BipartiteIndexedGraph(g::AbstractGraph)

Build a BipartiteIndexedGraph from any undirected, bipartite graph.

IndexedGraphs.nv_left — Function

nv_left(g::BipartiteIndexedGraph)

Return the number of vertices in the left block

IndexedGraphs.nv_right — Function

nv_right(g::BipartiteIndexedGraph)

Return the number of vertices in the right block

IndexedGraphs.Left — Type

Left

Singleton type used to represent a vertex belonging to the left block in a BipartiteGraphVertex

IndexedGraphs.Right — Type

Right

Singleton type used to represent a vertex belonging to the Right block in a BipartiteGraphVertex

IndexedGraphs.LeftorRight — Type

LeftorRight

LeftorRight = Union{Left, Right}

IndexedGraphs.BipartiteGraphVertex — Type

BipartiteGraphVertex

A BipartiteGraphVertex{LR<:LeftorRight,T<:Integer} represents a vertex in a bipartite graph.

PARAMETERS

LR – Either Left or Right

FIELDS

i – The index of the vertex within its block.

IndexedGraphs.vertex — Method

vertex(i::Integer, ::Type{<:LeftorRight})

Build a BipartiteGraphVertex

IndexedGraphs.vertex — Method

vertex(g::BipartiteIndexedGraph, i::Integer)

Build the BipartiteGraphVertex corresponding to linear index i. Throws an error if i is not in the range of vertices of g

IndexedGraphs.linearindex — Function

linearindex(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{LR}) where {LR<:LeftorRight}
linearindex(g::BipartiteIndexedGraph, i::Integer, ::Type{LR}) where LR<:LeftorRight

Return the linear index of a vertex, specified either by a BipartiteGraphVertex or by its index and block.

IndexedGraphs.vertices_left — Function

vertices_left(g::BipartiteIndexedGraph)

Return a lazy iterator to the vertices in the left block

IndexedGraphs.vertices_right — Function

vertices right(g::BipartiteIndexedGraph)

Return a lazy iterator to the vertices in the right block

IndexedGraphs.vertex_left — Function

vertex_left(g::BipartiteIndexedGraph, e::IndexedEdge)

Return the (in-block) index of the left-block vertex in e.

IndexedGraphs.vertex_right — Function

vertex_right(g::BipartiteIndexedGraph, e::IndexedEdge)

Return the (in-block) index of the right-block vertex in e.

IndexedGraphs.inedges — Method

inedges(g::BipartiteIndexedGraph, i::Integer)

Return a lazy iterator to the edges incident on a variable i specified by its linear index, with i as the destination.

IndexedGraphs.inedges — Method

inedges(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{<:LeftorRight})

Return a lazy iterator to the edges incident on a variable v specified by a BipartiteGraphVertex, with v as the destination.

IndexedGraphs.outedges — Method

outedges(g::BipartiteIndexedGraph, i::Integer)

Return a lazy iterator to the edges incident on a variable i specified by its linear index, with i as the source.

IndexedGraphs.outedges — Method

outedges(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{<:LeftorRight})

Return a lazy iterator to the edges incident on a variable v specified by a BipartiteGraphVertex, with v as the source.

Overrides from Graphs.jl

Graphs.inneighbors — Method

inneighbors(g::BipartiteIndexedGraph, i::Integer)

Return a lazy iterator to the neighbors of variable i specified by its linear index.

Graphs.inneighbors — Method

inneighbors(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{<:LeftorRight})

Return a lazy iterator to the neighbors of variable v specified by a BipartiteGraphVertex.

Graphs.outneighbors — Method

outneighbors(g::BipartiteIndexedGraph, i::Integer)

Return a lazy iterator to the neighbors of variable i specified by its linear index.

Graphs.outneighbors — Method

outneighbors(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{<:LeftorRight})

Return a lazy iterator to the neighbors of variable v specified by a BipartiteGraphVertex.

Graphs.LinAlg.adjacency_matrix — Function

adjacency_matrix(g::BipartiteIndexedGraph, T::DataType=Int)

Return the symmetric adjacency matrix of size nv(g) = nv_left(g) + nv_right(g) where no distinction is made between left and right nodes.