IndexedDiGraph

A type representing directed graphs.

Use this when you need to access only outedges and outneighbors. If you also need access to inedges and inneighbors, check out IndexedBiDiGraph.

IndexedGraphs.IndexedDiGraph — Type

IndexedDiGraph{T<:Integer} <: AbstractIndexedDiGraph{T}

A type representing a sparse directed graph with access only to outedges.

FIELDS

A – square matrix filled with NullNumbers. A[i,j] corresponds to an edge j=>i

IndexedGraphs.IndexedDiGraph — Method

IndexedDiGraph(A::AbstractMatrix)

Constructs a IndexedDiGraph from the adjacency matrix A.

IndexedDiGraph internally stores the transpose of A. To avoid overhead due to the transposition, use IndexedDiGraph(transpose(At)) where At is the transpose of A.