FactorGraph

A type representing Factor Graphs.

IndexedGraphs.FactorGraph — Type

FactorGraph{T<:Integer} <: AbstractIndexedGraph{T}

A type representing a sparse factor graph.

FIELDS

A – adjacency matrix filled with NullNumbers. Rows are factors, columns are variables

X – square matrix for efficient access by row. X[j,i] points to the index of element A[i,j] in A.nzval.

IndexedGraphs.FactorGraph — Method

FactorGraph(A::AbstractMatrix)

Construct a factor graph from adjacency matrix A with the convention that rows are factors, columns are variables

IndexedGraphs.FactorGraph — Method

FactorGraph(g::IndexedGraph)

Construct a factor graph whose factors are the pair-wise interactions encoded in g.

IndexedGraphs.get_edge — Method

get_edge(g::FactorGraph, v::Variable, f::Factor)
get_edge(g::FactorGraph, f::Factor, v::Variable)
get_edge(g::FactorGraph, id::Integer)

Get edge given variable and factor or given edge index

IndexedGraphs.bipartite_view — Function

bipartite_view(g::FactorGraph, T::DataType=Int)

Construct an undirected bipartite IndexedGraph where vertices are the concatenation of the variables and factors of the original FactorGraph. Edge indices are preserved in the bottom-left block.

Overrides from Graphs.jl

Graphs.neighbors — Method

neighbors(g::FactorGraph, v::Variable)

Return a lazy iterator to the neighbors of variable v

Graphs.neighbors — Method

neighbors(g::FactorGraph, f::Factor)

Return a lazy iterator to the neighbors of factor f

Graphs.edges — Method

edges(g::FactorGraph, v::Variable)

Return a lazy iterator to the edges incident to variable v

Graphs.edges — Method

edges(g::FactorGraph, f::Factor)

Return a lazy iterator to the edges incident to factor f

Graphs.LinAlg.adjacency_matrix — Function

adjacency_matrix(g::FactorGraph, T::DataType=Int)

Return the symmetric adjacency matrix of size nvariables(g) + nfactors(g) where no distinction is made between variable and factor nodes. Edge indices are preserved in the bottom-left block.