IndexedGraph

A type representing undirected graphs.

IndexedGraphs.IndexedGraph — Type

IndexedGraph{T<:Integer} <: AbstractIndexedGraph{T}

A type representing a sparse undirected graph.

FIELDS

A – square adjacency matrix. A[i,j] == A[j,i] contains the unique index associated to unidrected edge (i,j)

IndexedGraphs.IndexedGraph — Method

IndexedGraph(A::AbstractMatrix)

Construct an IndexedGraph from symmetric adjacency matrix A.

IndexedGraphs.inedges — Method

inedges(g::IndexedGraph, i::Integer)

Return a lazy iterators to the edges incident to i with i as the destination.

IndexedGraphs.outedges — Method

outedges(g::IndexedGraph, i::Integer)

Return a lazy iterators to the edges incident to i with i as the source.

IndexedGraphs.get_edge — Method

get_edge(g::IndexedGraph, src::Integer, dst::Integer)
get_edge(g::IndexedGraph, id::Integer)

Get edge given source and destination or given edge index.

Overrides from Graphs.jl

Graphs.edges — Method

edges(g::IndexedGraph, i::Integer)

Return a lazy iterators to the edges incident to i.

By default unordered edges sort source and destination nodes in increasing order. See outedges and inedges if you need otherwise.