Reference

IndexedGraphs.AbstractIndexedDiGraph — Type

AbstractIndexedDiGraph{T} <: AbstractIndexedGraph{T}

Abstract type for representing directed graphs.

IndexedGraphs.AbstractIndexedEdge — Type

AbstractIndexedEdge{T<:Integer} <: AbstractEdge{T}

Abstract type for indexed edge. AbstractIndexedEdge{T}s must have the following elements:

idx::T integer positive index

IndexedGraphs.AbstractIndexedGraph — Type

AbstractIndexedGraph{T} <: AbstractGraph{T}

An abstract type representing an indexed graph. AbstractIndexedGraphs must have the following elements:

A::SparseMatrixCSC adjacency matrix

IndexedGraphs.FactorGraph — Type

FactorGraph{T<:Integer} <: AbstractIndexedGraph{T}

A type representing a sparse factor graph.

FIELDS

A – adjacency matrix filled with NullNumbers. Rows are factors, columns are variables

X – square matrix for efficient access by row. X[j,i] points to the index of element A[i,j] in A.nzval.

IndexedGraphs.FactorGraphEdge — Type

FactorGraphEdge{T<:Integer} <: AbstractIndexedEdge{T}

Edge type for FactorGraphs.

Safe constructors

FactorGraphEdge(f::Factor, v::Variable, idx::Integer)
FactorGraphEdge(v::Variable, f::Factor, idx::Integer)

Example

FactorGraphEdge(Factor(2), Variable(4), 3)

Note

Specifying the variable first is also valid: FactorGraphEdge(v::Variable, f::Factor, idx::Integer)

IndexedGraphs.IndexedEdge — Type

IndexedEdge{T<:Integer} <: AbstractIndexedEdge{T}

Edge type for IndexedGraphs. Edge indices can be used to access edge properties stored in separate containers.

Graphs.LinAlg.adjacency_matrix — Function

Graphs.LinAlg.adjacency_matrix(g::FactorGraph, T::DataType=Int)

Return the symmetric adjacency matrix of size nvariables(g) + nfactors(g) where no distinction is made between variable and factor nodes. Edge indices are preserved in the bottom-left block.

IndexedGraphs.bipartite_view — Function

bipartite_view(g::FactorGraph, T::DataType=Int)

Construct an undirected bipartite SparseGraph where vertices are the concatenation of the variables and factors of the original FactorGraph. Edge indices are preserved in the bottom-left block.

IndexedGraphs.get_edge — Method

get_edge(g::AbstractIndexedDiGraph, src::Integer, dst::Integer)
get_edge(g::AbstractIndexedDiGraph, id::Integer)

Get edge given source and destination or given edge index.

IndexedGraphs.get_edge — Method

get_edge(g::FactorGraph, v::Variable, f::Factor)
get_edge(g::FactorGraph, f::Factor, v::Variable)
get_edge(g::FactorGraph, id::Integer)

Get edge given variable and factor or given edge index

IndexedGraphs.get_edge — Method

get_edge(g::IndexedGraph, src::Integer, dst::Integer)
get_edge(g::IndexedGraph, id::Integer)

Get edge given source and destination or given edge index.

IndexedGraphs.inedges — Method

inedges(g::AbstractIndexedDiGraph, i::Integer)

Return a lazy iterator to the edges ingoing to node i in g.

IndexedGraphs.outedges — Method

outedges(g::AbstractIndexedDiGraph, i::Integer)

Return a lazy iterator to the edges outgoing from node i in g.