FactorGraph

A type representing Factor Graphs.

IndexedGraphs.FactorGraph — Type

FactorGraph{T<:Integer} <: AbstractIndexedGraph{T}

A type representing a sparse factor graph.

FIELDS

A – adjacency matrix filled with NullNumbers. Rows are factors, columns are variables

X – square matrix for efficient access by row. X[j,i] points to the index of element A[i,j] in A.nzval.

IndexedGraphs.FactorGraph — Method

FactorGraph(A::AbstractMatrix)

Construct a factor graph from adjacency matrix A with the convention that rows are factors, columns are variables

IndexedGraphs.get_edge — Method

get_edge(g::FactorGraph, v::Variable, f::Factor)
get_edge(g::FactorGraph, f::Factor, v::Variable)
get_edge(g::FactorGraph, id::Integer)

Get edge given variable and factor or given edge index

IndexedGraphs.bipartite_view — Function

bipartite_view(g::FactorGraph, T::DataType=Int)

Construct an undirected bipartite IndexedGraph where vertices are the concatenation of the variables and factors of the original FactorGraph. Edge indices are preserved in the bottom-left block.

Overrides from Graphs.jl

Missing docstring.

Missing docstring for neighbors(g::FactorGraph, v::Variable). Check Documenter's build log for details.

Missing docstring.

Missing docstring for neighbors(g::FactorGraph, f::Factor). Check Documenter's build log for details.

Missing docstring.

Missing docstring for edges(g::FactorGraph, v::Variable). Check Documenter's build log for details.

Missing docstring.

Missing docstring for edges(g::FactorGraph, f::Factor). Check Documenter's build log for details.

Missing docstring.

Missing docstring for adjacency_matrix(g::FactorGraph, T::DataType=Int). Check Documenter's build log for details.