Reference

IndexedGraphs.LeftorRight — Type

LeftorRight

LeftorRight = Union{Left, Right}

IndexedGraphs.AbstractIndexedDiGraph — Type

AbstractIndexedDiGraph{T} <: AbstractIndexedGraph{T}

Abstract type for representing directed graphs.

IndexedGraphs.AbstractIndexedEdge — Type

AbstractIndexedEdge{T<:Integer} <: AbstractEdge{T}

Abstract type for indexed edge. AbstractIndexedEdge{T}s must have the following elements:

idx::T integer positive index

IndexedGraphs.AbstractIndexedGraph — Type

AbstractIndexedGraph{T} <: AbstractGraph{T}

An abstract type representing an indexed graph. AbstractIndexedGraphs must have the following elements:

A::SparseMatrixCSC adjacency matrix

IndexedGraphs.BipartiteGraphVertex — Type

BipartiteGraphVertex

A BipartiteGraphVertex{LR<:LeftorRight,T<:Integer} represents a vertex in a bipartite graph.

PARAMETERS

LR – Either Left or Right

FIELDS

i – The index of the vertex within its block.

IndexedGraphs.BipartiteIndexedGraph — Method

BipartiteIndexedGraph(g::AbstractGraph)

Build a BipartiteIndexedGraph from any undirected, bipartite graph.

IndexedGraphs.BipartiteIndexedGraph — Method

BipartiteIndexedGraph(A::AbstractMatrix)

Construct a BipartiteIndexedGraph from adjacency matrix A with the convention that rows are vertices belonging to the left block, columns to the right block

IndexedGraphs.BipartiteIndexedGraph — Type

BipartiteIndexedGraph{T<:Integer} <: AbstractIndexedGraph{T}

A type representing a sparse, undirected bipartite graph.

FIELDS

A – adjacency matrix filled with NullNumbers. Rows are vertices belonging to the left block, columns to the right block

X – square matrix for efficient access by row. X[j,i] points to the index of element A[i,j] in A.nzval.

IndexedGraphs.IndexedBiDiGraph — Method

IndexedBiDiGraph(A::AbstractMatrix)

Construct an IndexedBiDiGraph from the adjacency matrix A.

IndexedBiDiGraph internally stores the transpose of A. To avoid overhead due to the transposition, use IndexedBiDiGraph(transpose(At)) where At is the transpose of A.

IndexedGraphs.IndexedBiDiGraph — Method

IndexedBiDiGraph(A::AbstractSimpleGraph)

Construct an IndexedBiDiGraph from any AbstractSimpleGraph (Graphs.jl), directed or otherwise.

IndexedGraphs.IndexedBiDiGraph — Type

IndexedBiDiGraph{T<:Integer} <: AbstractIndexedDiGraph{T}

A type representing a sparse directed graph with access to both outedges and inedges.

FIELDS

A – square matrix filled with NullNumbers. A[i,j] corresponds to edge j=>i.

X – square matrix for efficient access by row. X[j,i] points to the index of element A[i,j] in A.nzval.

IndexedGraphs.IndexedDiGraph — Method

IndexedDiGraph(A::AbstractMatrix)

Constructs a IndexedDiGraph from the adjacency matrix A.

IndexedDiGraph internally stores the transpose of A. To avoid overhead due to the transposition, use IndexedDiGraph(transpose(At)) where At is the transpose of A.

IndexedGraphs.IndexedDiGraph — Method

IndexedDiGraph(A::AbstractSimpleGraph)

Construct an IndexedDiGraph from any AbstractSimpleGraph (Graphs.jl), directed or otherwise.

IndexedGraphs.IndexedDiGraph — Type

IndexedDiGraph{T<:Integer} <: AbstractIndexedDiGraph{T}

A type representing a sparse directed graph with access only to outedges.

FIELDS

A – square matrix filled with NullNumbers. A[i,j] corresponds to an edge j=>i

IndexedGraphs.IndexedEdge — Type

IndexedEdge{T<:Integer} <: AbstractIndexedEdge{T}

Edge type for IndexedGraphs. Edge indices can be used to access edge properties stored in separate containers.

IndexedGraphs.IndexedGraph — Method

IndexedGraph(A::AbstractMatrix)

Construct an IndexedGraph from symmetric adjacency matrix A.

IndexedGraphs.IndexedGraph — Method

IndexedGraph(A::SimpleGraph)

Construct an IndexedGraph from undirected SimpleGraph (Graphs.jl).

IndexedGraphs.IndexedGraph — Type

IndexedGraph{T<:Integer} <: AbstractIndexedGraph{T}

A type representing a sparse undirected graph.

FIELDS

A – square adjacency matrix. A[i,j] == A[j,i] contains the unique index associated to unidrected edge (i,j)

IndexedGraphs.Left — Type

Left

Singleton type used to represent a vertex belonging to the left block in a BipartiteGraphVertex

IndexedGraphs.Right — Type

Right

Singleton type used to represent a vertex belonging to the Right block in a BipartiteGraphVertex

Graphs.LinAlg.adjacency_matrix — Function

adjacency_matrix(g::BipartiteIndexedGraph, T::DataType=Int)

Return the symmetric adjacency matrix of size nv(g) = nv_left(g) + nv_right(g) where no distinction is made between left and right nodes.

Graphs.edges — Method

edges(g::IndexedGraph, i::Integer)

Return a lazy iterators to the edges incident to i.

By default unordered edges sort source and destination nodes in increasing order. See outedges and inedges if you need otherwise.

Graphs.inneighbors — Method

inneighbors(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{<:LeftorRight})

Return a lazy iterator to the neighbors of variable v specified by a BipartiteGraphVertex.

Graphs.inneighbors — Method

inneighbors(g::BipartiteIndexedGraph, i::Integer)

Return a lazy iterator to the neighbors of variable i specified by its linear index.

Graphs.outneighbors — Method

outneighbors(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{<:LeftorRight})

Return a lazy iterator to the neighbors of variable v specified by a BipartiteGraphVertex.

Graphs.outneighbors — Method

outneighbors(g::BipartiteIndexedGraph, i::Integer)

Return a lazy iterator to the neighbors of variable i specified by its linear index.

IndexedGraphs.bidirected_with_mappings — Method

bidirected_with_mappings(g::IndexedGraph) -> (gdir, dir2undir, undir2dir)

Construct an IndexedBiDiGraph gdir from an IndexedGraph g by building two directed edges per every undirected edge in g. In addition, return two vectors containing mappings from the undirected edges of g to the corresponding directed edges of gdir.

OUTPUT

gdir – The directed graph

dir2undir – A vector of integers mapping the indices of the directed edges of gdir to the corresponding undirected edges of g

undir2dir – A vector of vectors with two integers each mapping the indices of the undirected edges of g to the two corresponding directed edges of gdir

IndexedGraphs.get_edge — Method

get_edge(g::AbstractIndexedDiGraph, src::Integer, dst::Integer)
get_edge(g::AbstractIndexedDiGraph, id::Integer)

Get edge given source and destination or given edge index.

IndexedGraphs.get_edge — Method

get_edge(g::IndexedGraph, src::Integer, dst::Integer)
get_edge(g::IndexedGraph, id::Integer)

Get edge given source and destination or given edge index.

IndexedGraphs.inedges — Method

inedges(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{<:LeftorRight})

Return a lazy iterator to the edges incident on a variable v specified by a BipartiteGraphVertex, with v as the destination.

IndexedGraphs.inedges — Method

inedges(g::BipartiteIndexedGraph, i::Integer)

Return a lazy iterator to the edges incident on a variable i specified by its linear index, with i as the destination.

IndexedGraphs.inedges — Method

inedges(g::AbstractIndexedBiDiGraph, i::Integer)

Return a lazy iterator to the edges ingoing to node i in g.

IndexedGraphs.inedges — Method

inedges(g::IndexedGraph, i::Integer)

Return a lazy iterators to the edges incident to i with i as the destination.

IndexedGraphs.linearindex — Method

linearindex(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{LR}) where {LR<:LeftorRight}
linearindex(g::BipartiteIndexedGraph, i::Integer, ::Type{LR}) where LR<:LeftorRight

Return the linear index of a vertex, specified either by a BipartiteGraphVertex or by its index and block.

IndexedGraphs.nv_left — Method

nv_left(g::BipartiteIndexedGraph)

Return the number of vertices in the left block

IndexedGraphs.nv_right — Method

nv_right(g::BipartiteIndexedGraph)

Return the number of vertices in the right block

IndexedGraphs.outedges — Method

outedges(g::AbstractIndexedDiGraph, i::Integer)

Return a lazy iterator to the edges outgoing from node i in g.

IndexedGraphs.outedges — Method

outedges(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{<:LeftorRight})

Return a lazy iterator to the edges incident on a variable v specified by a BipartiteGraphVertex, with v as the source.

IndexedGraphs.outedges — Method

outedges(g::BipartiteIndexedGraph, i::Integer)

Return a lazy iterator to the edges incident on a variable i specified by its linear index, with i as the source.

IndexedGraphs.outedges — Method

outedges(g::IndexedGraph, i::Integer)

Return a lazy iterators to the edges incident to i with i as the source.

IndexedGraphs.vertex — Method

vertex(g::BipartiteIndexedGraph, i::Integer)

Build the BipartiteGraphVertex corresponding to linear index i. Throws an error if i is not in the range of vertices of g

IndexedGraphs.vertex — Method

vertex(i::Integer, ::Type{<:LeftorRight})

Build a BipartiteGraphVertex

IndexedGraphs.vertex_left — Method

vertex_left(g::BipartiteIndexedGraph, e::IndexedEdge)

Return the (in-block) index of the left-block vertex in e.

IndexedGraphs.vertex_right — Method

vertex_right(g::BipartiteIndexedGraph, e::IndexedEdge)

Return the (in-block) index of the right-block vertex in e.

IndexedGraphs.vertices_left — Method

vertices_left(g::BipartiteIndexedGraph)

Return a lazy iterator to the vertices in the left block

IndexedGraphs.vertices_right — Method

vertices right(g::BipartiteIndexedGraph)

Return a lazy iterator to the vertices in the right block

LinearAlgebra.issymmetric — Method

issymmetric(g::IndexedBiDiGraph) -> Bool

Test whether a directed graph is symmetric, i.e. for each directed edge i=>j there also exists the edge j=>i