Reference

IndexedGraphs.AbstractIndexedDiGraph — Type

AbstractIndexedDiGraph{T} <: AbstractIndexedGraph{T}

Abstract type for representing directed graphs.

IndexedGraphs.AbstractIndexedEdge — Type

AbstractIndexedEdge{T<:Integer} <: AbstractEdge{T}

Abstract type for indexed edge. AbstractIndexedEdge{T}s must have the following elements:

idx::T integer positive index

IndexedGraphs.AbstractIndexedGraph — Type

AbstractIndexedGraph{T} <: AbstractGraph{T}

An abstract type representing an indexed graph. AbstractIndexedGraphs must have the following elements:

A::SparseMatrixCSC adjacency matrix

IndexedGraphs.IndexedBiDiGraph — Method

IndexedBiDiGraph(A::AbstractSimpleGraph)

Construct an IndexedBiDiGraph from any AbstractSimpleGraph (Graphs.jl), directed or otherwise.

IndexedGraphs.IndexedDiGraph — Method

IndexedDiGraph(A::AbstractSimpleGraph)

Construct an IndexedDiGraph from any AbstractSimpleGraph (Graphs.jl), directed or otherwise.

IndexedGraphs.IndexedEdge — Type

IndexedEdge{T<:Integer} <: AbstractIndexedEdge{T}

Edge type for IndexedGraphs. Edge indices can be used to access edge properties stored in separate containers.

IndexedGraphs.IndexedGraph — Method

IndexedGraph(A::SimpleGraph)

Construct an IndexedGraph from undirected SimpleGraph (Graphs.jl).

IndexedGraphs.get_edge — Method

get_edge(g::AbstractIndexedDiGraph, src::Integer, dst::Integer)
get_edge(g::AbstractIndexedDiGraph, id::Integer)

Get edge given source and destination or given edge index.

IndexedGraphs.inedges — Method

inedges(g::AbstractIndexedBiDiGraph, i::Integer)

Return a lazy iterator to the edges ingoing to node i in g.

IndexedGraphs.linearindex — Method

linearindex(g::BipartiteIndexedGraph, v::BipartiteGraphVertex{LR}) where {LR<:LeftorRight}
linearindex(g::BipartiteIndexedGraph, i::Integer, ::Type{LR}) where LR<:LeftorRight

Return the linear index of a vertex, specified either by a BipartiteGraphVertex or by its index and block.

IndexedGraphs.nv_left — Method

nv_left(g::BipartiteIndexedGraph)

Return the number of vertices in the left block

IndexedGraphs.nv_right — Method

nv_right(g::BipartiteIndexedGraph)

Return the number of vertices in the right block

IndexedGraphs.outedges — Method

outedges(g::AbstractIndexedDiGraph, i::Integer)

Return a lazy iterator to the edges outgoing from node i in g.

IndexedGraphs.vertex — Method

vertex(i::Integer, ::Type{<:LeftorRight})

Build a BipartiteGraphVertex

IndexedGraphs.vertex_left — Method

vertex_left(g::BipartiteIndexedGraph, e::IndexedEdge)

Return the (in-block) index of the left-block vertex in e.

IndexedGraphs.vertex_right — Method

vertex_right(g::BipartiteIndexedGraph, e::IndexedEdge)

Return the (in-block) index of the right-block vertex in e.

IndexedGraphs.vertices_left — Method

vertices_left(g::BipartiteGraphVertex)

Return a lazy iterator to the vertices in the left block

IndexedGraphs.vertices_right — Method

vertices right(g::BipartiteGraphVertex)

Return a lazy iterator to the vertices in the right block