Reference

IndexedGraphs.AbstractIndexedDiGraph — Type

AbstractIndexedDiGraph{T} <: AbstractIndexedGraph{T}

Abstract type for representing directed graphs.

IndexedGraphs.AbstractIndexedEdge — Type

AbstractIndexedEdge{T<:Integer} <: AbstractEdge{T}

Abstract type for indexed edge. AbstractIndexedEdge{T}s must have the following elements:

idx::T integer positive index

IndexedGraphs.AbstractIndexedGraph — Type

AbstractIndexedGraph{T} <: AbstractGraph{T}

An abstract type representing an indexed graph. AbstractIndexedGraphs must have the following elements:

A::SparseMatrixCSC adjacency matrix

IndexedGraphs.Factor — Type

Factor{T<:Integer}

Wraps an index to specify that it is the index of a factor. See e.g. neighbors(g::FactorGraph, f::Factor)

IndexedGraphs.FactorGraphEdge — Type

FactorGraphEdge{T<:Integer} <: AbstractIndexedEdge{T}

Edge type for FactorGraphs.

Safe constructors

FactorGraphEdge(f::Factor, v::Variable, idx::Integer)
FactorGraphEdge(v::Variable, f::Factor, idx::Integer)

Example:

    FactorGraphEdge(Factor(2), Variable(4), 3)

IndexedGraphs.IndexedEdge — Type

IndexedEdge{T<:Integer} <: AbstractIndexedEdge{T}

Edge type for IndexedGraphs. Edge indices can be used to access edge properties stored in separate containers.

IndexedGraphs.Variable — Type

Variable{T<:Integer}

Wraps an index to specify that it is the index of a variable. See e.g. neighbors(g::FactorGraph, v::Variable)

Graphs.LinAlg.adjacency_matrix — Function

adjacency_matrix(g::FactorGraph, T::DataType=Int)

Return the symmetric adjacency matrix of size nvariables(g) + nfactors(g) where no distinction is made between variable and factor nodes. Edge indices are preserved in the bottom-left block.

Graphs.edges — Method

neighbors(g::FactorGraph, f::Factor)

Return a lazy iterator to the edges incident to factor f

Graphs.edges — Method

edges(g::FactorGraph, v::Variable)

Return a lazy iterator to the edges incident to variable v

Graphs.edges — Method

edges(g::IndexedGraph, i::Integer)

Return a lazy iterators to the edges incident to i.

By default unordered edges sort source and destination nodes in increasing order. See outedges and inedges if you need otherwise.

Graphs.neighbors — Method

neighbors(g::FactorGraph, f::Factor)

Return a lazy iterator to the neighbors of factor f

Graphs.neighbors — Method

neighbors(g::FactorGraph, v::Variable)

Return a lazy iterator to the neighbors of variable v

IndexedGraphs.get_edge — Method

get_edge(g::AbstractIndexedDiGraph, src::Integer, dst::Integer)
get_edge(g::AbstractIndexedDiGraph, id::Integer)

Get edge given source and destination or given edge index.

IndexedGraphs.inedges — Method

inedges(g::AbstractIndexedBiDiGraph, i::Integer)

Return a lazy iterator to the edges ingoing to node i in g.

IndexedGraphs.inedges — Method

inedges(g::IndexedGraph, i::Integer)

Return a lazy iterators to the edges incident to i with i as the destination.

IndexedGraphs.outedges — Method

outedges(g::AbstractIndexedDiGraph, i::Integer)

Return a lazy iterator to the edges outgoing from node i in g.

IndexedGraphs.outedges — Method

outedges(g::IndexedGraph, i::Integer)

Return a lazy iterators to the edges incident to i with i as the source.